Fixed GC in atom allocations.

+++ interpreter/alloc.h
@@ -1,5 +1,5 @@
 /*
- * tré - Copyright (c) 2005-2009 Sven Michael Klose <pixel@copei.de>
+ * tré – Copyright (c) 2005–2009,2013 Sven Michael Klose <pixel@copei.de>
  */
 
 #ifndef TRE_ALLOC_H
@@ -17,6 +17,6 @@ extern void   trealloc_free (void *);
 
 extern void * trealloc_item (void * handle);
 extern void   trealloc_free_item (void * handle, void * item);
-extern void * trealloc_item_init (void * map, ulong num_items, ulong item_size);
+extern void * trealloc_item_init (void * map, size_t num_items, size_t item_size);
 
 #endif
